chore: migrate Rstest to Rstack CLI
