+++ e2e/tsconfig.json
@@ -6,7 +6,7 @@
   },
   "include": [
     "helper",
-    "rstest.config.ts",
+    "rstack.config.ts",
     "cases/**/*.test.ts",
     "cases/**/*.config.ts",
     "cases/javascript-api/server-custom-message/src/**/*.ts"
