+++ package.json
@@ -17,12 +17,11 @@
     "prebundle": "pnpm --parallel --filter \"./packages/*\" run prebundle",
     "prepare": "simple-git-hooks && node --run prebundle",
     "sort-package-json": "pnpx sort-package-json \"./package.json\" \"packages/*/package.json\"",
-    "test": "rstest",
-    "test:watch": "rstest watch"
+    "test": "rs test",
+    "test:watch": "rs test watch"
   },
   "devDependencies": {
     "@rstest/adapter-rslib": "catalog:",
-    "@rstest/core": "catalog:",
     "@scripts/config": "workspace:*",
     "@scripts/test-helper": "workspace:*",
     "@types/node": "catalog:",
