+++ packages/plugin-vue/tests/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "extends": "@scripts/config/tsconfig",
   "compilerOptions": {
-    "types": ["@rstest/core/globals"]
+    "types": ["rstack/test/globals"]
   },
   "include": ["./**/*.ts"]
 }
